+++ webrtc/modules/audio_conference_mixer/source/audio_conference_mixer_impl.cc
@@ -197,8 +197,8 @@ int32_t AudioConferenceMixerImpl::ChangeUniqueId(const int32_t id) {
 }
 
 // Process should be called every kProcessPeriodicityInMs ms
-int32_t AudioConferenceMixerImpl::TimeUntilNextProcess() {
-    int32_t timeUntilNextProcess = 0;
+int64_t AudioConferenceMixerImpl::TimeUntilNextProcess() {
+    int64_t timeUntilNextProcess = 0;
     CriticalSectionScoped cs(_crit.get());
     if(_timeScheduler.TimeToNextUpdate(timeUntilNextProcess) != 0) {
         WEBRTC_TRACE(kTraceError, kTraceAudioMixerServer, _id,
