+++ webrtc/modules/audio_device/audio_device_impl.cc
@@ -480,7 +480,7 @@ int32_t AudioDeviceModuleImpl::ChangeUniqueId(const int32_t id)
 //  to call Process().
 // ----------------------------------------------------------------------------
 
-int32_t AudioDeviceModuleImpl::TimeUntilNextProcess()
+int64_t AudioDeviceModuleImpl::TimeUntilNextProcess()
 {
     uint32_t now = AudioDeviceUtility::GetTimeInMS();
     int32_t deltaProcess = kAdmMaxIdleTimeProcess - (now - _lastProcessTime);
