+++ webrtc/modules/media_file/source/media_file_impl.h
@@ -28,7 +28,7 @@ public:
 
     virtual int32_t ChangeUniqueId(const int32_t id) OVERRIDE;
     virtual int32_t Process() OVERRIDE;
-    virtual int32_t TimeUntilNextProcess() OVERRIDE;
+    virtual int64_t TimeUntilNextProcess() OVERRIDE;
 
     // MediaFile functions
     virtual int32_t PlayoutAudioData(int8_t* audioBuffer,
