+++ webrtc/modules/video_render/include/video_render.h
@@ -59,7 +59,7 @@ public:
      */
     virtual int32_t ChangeUniqueId(const int32_t id) OVERRIDE = 0;
 
-    virtual int32_t TimeUntilNextProcess() OVERRIDE = 0;
+    virtual int64_t TimeUntilNextProcess() OVERRIDE = 0;
     virtual int32_t Process() OVERRIDE = 0;
 
     /**************************************************************************
