+++ webrtc/modules/video_render/video_render_impl.cc
@@ -129,7 +129,7 @@ int32_t ModuleVideoRenderImpl::ChangeUniqueId(const int32_t id)
     return 0;
 }
 
-int32_t ModuleVideoRenderImpl::TimeUntilNextProcess()
+int64_t ModuleVideoRenderImpl::TimeUntilNextProcess()
 {
     // Not used
     return 50;
